substack-archiver/pyproject.toml
[project]
name = "substack-playwright"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "fastmcp>=2.14.4",
    "pillow>=12.1.0",
    "playwright>=1.57.0",
    "sniffio>=1.3.1",
    "zai-sdk>=0.2.1",
]
